requirements.txt
gymnasium==1.1.1
imageio==2.33.1
ipython==8.12.3
matplotlib==3.10.1
numpy==2.2.5
pandas==2.2.3
torch==2.5.1
~orch==2.6.0
imageio[ffmpeg]
hydra-core==1.3.2
mujoco 
glfw